api update

+++ database/migrations/2019_09_29_200423_create_media_table.php
@@ -19,7 +19,7 @@ class CreateMediaTable extends Migration
             $table->text('short_description');
             $table->text('description');
             $table->string('trailer_url');
-            $table->double('imdb_rating', 4, 2)->nullable();
+            $table->decimal('imdb_rating', 4, 2)->nullable();
             $table->timestamps();
         });
     }
